Spread damage when attacking human-shaped
