+++ src/creature.h
@@ -239,8 +239,7 @@ class Creature
 
         // Makes a ranged projectile attack against the creature
         // Sets relevant values in `attack`.
-        virtual void deal_projectile_attack( Creature *source, dealt_projectile_attack &attack,
-                                             bool print_messages = true );
+        virtual void deal_projectile_attack( Creature *source, dealt_projectile_attack &attack );
 
         /**
          * Deals the damage via an attack. Allows armor mitigation etc.
