+++ src/monster.h
@@ -308,8 +308,7 @@ class monster : public Creature
         void melee_attack( Creature &target );
         void melee_attack( Creature &target, float accuracy );
         void melee_attack( Creature &p, bool ) = delete;
-        void deal_projectile_attack( Creature *source, dealt_projectile_attack &attack,
-                                     bool print_messages = true ) override;
+        void deal_projectile_attack( Creature *source, dealt_projectile_attack &attack ) override;
         void deal_damage_handle_type( const damage_unit &du, bodypart_id bp, int &damage,
                                       int &pain ) override;
         void apply_damage( Creature *source, bodypart_id bp, int dam,
